refactor: style system revision v4
I kept testing the style system in various situations and wasn't satisfied with the way it worked, ran into issues with tailwind, or it wouldn't correctly style for certain states. So after a few revisions of different ways to build the system, it seems I finally found a solution that satisfies all of those restrictions/requirements.

- Revised style system and related files, presets, and components that use it.